% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geohashr.R
\name{geohash_decode}
\alias{geohash_decode}
\title{Decode geohash.}
\usage{
geohash_decode(hash)
}
\arguments{
\item{hash}{character vector, geohashes}
}
\value{
data.frame with columns lat and lng, corresponding to \code{hash}.
}
\description{
Decode geohash.
}
\examples{
geohash_decode(c("7zzz", "9hmmy"))
}

