% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geohashr.R
\name{geohash_encode}
\alias{geohash_encode}
\title{Encode geohash.}
\usage{
geohash_encode(lat, lng, precision)
}
\arguments{
\item{lat}{numeric vector, latitudes}

\item{lng}{numeric vector, longitudes}

\item{precision}{integer, precision}
}
\value{
Geohash, as character vector.
}
\description{
Encode geohash.
}
\examples{
new_geohash <- geohash_encode(c(0, 24), c(0, -25), 12)
}

