% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geohashr.R
\name{geohash_encode_sp}
\alias{geohash_encode_sp}
\title{Encode geohash from sp.}
\usage{
geohash_encode_sp(points_obj, precision)
}
\arguments{
\item{points_obj}{\code{SpatialPoints} or \code{SpatialPointsDataFrame}
object}

\item{precision}{integer, precision}
}
\value{
Geohash, as character vector.
}
\description{
Encode geohash from sp.
}
\examples{
library(sp)
geohash_encode_sp(SpatialPoints(cbind((-20):(-25), 0:5)), precision = 12)
}

