% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geohashr.R
\name{sp_geohash_decode}
\alias{sp_geohash_decode}
\title{Decode geohash to sp.}
\usage{
sp_geohash_decode(hash)
}
\arguments{
\item{hash}{character vector, geohashes}
}
\value{
\code{SpatialPoints} object corresponding to \code{hash}.
}
\description{
Decode geohash to sp.
}
\examples{
library(sp)
sp_geohash_decode(c("7zzz", "9hmmy"))
}

